'amd64' is also ok for sparse-llvm
For sparse-llvm, the only layouts are for x86 architecture,
either 32 or 64bit. So, it's not built for other archs.

This is done by parsing the output of 'uname -m' but only
'i[3456]86' or 'x86' are accepted as pattern while 'amd64'
is also used.

Change the pattern to also accept 'amd64'.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>