comparison: improve handling of "if (foo + 1 > limit) "
We could have a condition "if (foo < limit - 1) " or
"if (foo + 1 < limit)".  It's basically the same.  But then it's hard to
compare because we could compare "offset < limit - 1" but do
"array[offset + 1]".  So it becomes an issue that the comparison and use
don't match exactly.

Let's work around that by always writing the comparison as a minus.  So
we move the "+ 1" to the other side of the equation to become "- 1" with
algebra.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>