implied: use assume() even if there are no implications
I'm basically using this to take "array[offset];" and I say let's assume
that "offset" is out of bounds.  Then I check that it doesn't cause an
impossible condition.  Even if there are no "implications" (as in we
haven't encountered that exact condition before) the new stree can still
be useful.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>