extra: handle "if (foo & 0x1) {"
The following code is a pretty common idiom:

	if (!(offset & 0x1))
		array[offset + 1] = foo;

Say array[] has an even number of elements and we know that offset is
less than array size.  That means the last element is an odd number so we
can test for it and know that we're within bounds.

Anyway, the point is that it's a common idiom and we can test for it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>