apply_ctype: reverse the order of arguments
apply_ctype()'s argument order confuse me endlessly as I'm much more
used to have the destination first and the source next (the so called
'assignment order' used for assignments but also in memcpy() and in
many sparse or library functions).

So, change the argument order of apply_ctype() to mimic the order
of memcpy()/assignment, to hopefully reduce my confusion.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>