fix: warn on typedef redefinition
check_duplicates() verifies that symbols are not redefined and
warns if they are.

However, this function is only called at evaluation-time and
then only for symbols corresponding to objects and functions.
So, typedefs can be redefined without any kind of diagnostic.

Fix this by calling check_duplicates() at parsing time on typedefs.

Note: this is C11's semantic or GCC's C89/C99 in non-pedantic mode.

Reported-by: Matthew Wilcox <willy@infradead.org>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>