llvm: normalize sparse-llvm-dis' output
Depending on some combination of host architecture/LLVM's version/...
the output of llvm-dis may or may not contains the target layout.
Worse it may also contains the target triple that was used to
compile sparse-llvm.

This is a problem for sparse-llvm-dis (which unique purpose is to test
the generated LLVM bytecode) given that the test-system doesn't offer
much flexibility: it only known about line patterns or the whole output
without having the possibility to ignore some lines.

Fix this by filtering out lines begining by 'target ' in the output of
sparse-llvm-dis.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>