avoid deadborn loads & stores
During linearization, most kinds of instruction are not generated
if there is no active BB (which mean that these instructions can
never be executed).

However, loads & stores are generated anyway. These dead loads
and stores will then need to be removed which is a bit tricky:
- memops are special and more complex to be removed than
  instructions like 'add' and such.
- these instructions exist in 'phantom basic blocks': a BB
  which has ia null bb->ep and which doesn't belong to then
  entrypoint's ep->bbs. Such blocks are considered as removed
  and are never scanned or anything.

Keep things simple and avoid to generate memops in inactive
basic blocks.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>