return_to_param: fix a crashing bug
This started happening when I upgraded to the latest Sparse.

So I'm not totally sure why I did this HACK ALERT...  And I also can't
swear that this fix is correct because the test cases aren't there...

But what happens is that we use the left_sym, because it gets passed to
smatch_extra mod expression.  So the hack says we don't use it but that's
wrong.  Then check_rosenberg.c hooks into smatch_extra and looks up the
type of the symbol which doesn't work since it's an expression pointer.
And it cause a crash.

I think this fix will work 99 times out of a hundred and it's probably
harmless the remaining time.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>