var_sym: tweak expr_to_vsl()
I'm not sure what's going on here exactly....

I changed the first hunk by removing the EXPR_DEREF from the condition.
Why did I do that?  Possibly because DEREF has an expr->deref instead of
a ->unop?  I don't know.  It's possible that it has to do with the
EXPR_SYMBOL thing?

But anyway, it meant that smatch_extra.c couldn't save the values for
array expressions like "if (foo[x] != 3) { ...".  So I added EXPR_DEREF
handling back later in the bottom of the function.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>