testsuite: fix problem with double-escaping in patterns
Since the patterns in the testcases are evaluated in the shell
script, the backslash used to escape characters special to the
pattern need itself to be escaped. Theer is a few cases where
it wasn't done so, partly because 'format -l' gave a single
escape in its template.

Fix all occurences neededing this double-escape as well as the
'format -l' template.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>