ranges: fix a leak
smatch_ranges.c has always been horribly leaky.  Basically, the only time
range lists are freed are when they are part of a state.

That hasn't been a huge issue, but always Smatch grows and uses more
resources and we have to be more careful.  There are a few massive defines
int the kernel and it's comes out to a single constant when you compile the
code but it has to parsed and maybe parsed several times by smatch and each
time we do math we allocate another range_list.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>