implied: introduce impossible_assumption()
This is to deal with the situation where you have an impossible path were
foo == 42, but you merge it with a possible path where foo == 41.  Now say
you have read from "array[foo]" and 41 is valid and 42 is invalid.  Then
it's useful to be able to say is "foo == 42" really possible?

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>