user_data: reading from an array doesn't necessarily give user data
Say you have a user defined offset into a kernel controlled array.  Smatch
was saying that because the offset was user controlled and the result was
unknown that meant that the result was an entirely unknown user controlled
value.

This silences that.  It's a big hammer approach, but it probably doesn't
silence any real bugs.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>