alloc: check if size is too big
The allocate functions can take an extra size in arguments
used to allocate some variable extta space at the end of the
allocated structure. In the common case this extra size is zero
and if not it should be relatively small. In all cases the total
size must be smaller than the 'chunking factor' (the size of the
block ued t do these allocations).

However, this total size is not tested and all kinds of interesting
failures can be produced if it is too big.

Fix this by adding a test and die in case of failure.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>