optim: fix REPEAT_CFG_CLEANUP
The REPEAT_... flags are used as bitfield but the last one
introduced (by myself): REPEAT_CFG_CLEANUP was wrongly defined
to the next value instead of the next bit.

Fix this by defining it to the right value.
Also change all the definition to use '1 << ...' to clearly
show we expect uniq bit values here.

Fixes: 917d37ad7681c88fc4d8df484f7b10d550190df8
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>