taint: let check_access() warn just once
The function check_access() verify that memory accesses are not
done out of their bounds. This is good.

However, this function is called at each run of simplify_loads()
which means that the same warning can be given multiple times
which is annoying.

Fix this by using the newly added 'tainted' field to not warn
a second time on this instruction.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>