cleanup deadborn phi-sources
In simplify_loads(), phisrcs are created in find_dominating_parents()
and are then supposed to be used in rewrite_load_instruction().

However, it may happen (quite often) that find_dominating_parents()
find a dominator for one of the branch, create a phi-source for it,
record it's usage and then doesn't find a dominator in one of other
parent branches. In this case, the function returns early and the
created phisrcs are simply ignored. These phisrcs can't be simplified
away as dead instructions because they still have their usage recorded.

Fix this by explicitly remove these ignored phisrcs.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>