ssa: phi worklist
This patch optimize the very simple implementation of the
phi-node renaming at the end of the SSA conversion.

It avoids the need to rescan the whole function to find the phi-nodes
by using a worklist to put the phi-nodes during the renaming
of non-phi nodes instructions.

This optimization avoids O(n^2) behaviour in some pathological cases.

Note: A lot of optimizations can be done for the renaming.
      For the moment, things are kept as simplest as possible,
      the goal being to have correctness first.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>