math: introduce a new cache of recently used constants
The issue here is that the kernel has several macros which expand to pages
and pages of calculations.  We end up calculating them over and over which
uses a ton of memory.  So parsing some files ended up using like over 50GB
of RAM.

When we introduce this cache it makes a huge improvement.

I'd like to cache implied range lists as well, but I was nervous about that
because the same expression can have multiple values depending on when it's
parsed.  The obvious one here is loop iterators but there maybe are others
as well, especially if we're dealing with fake strees etc.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>