add TODO list.
The Documentation directory contains a 'project ideas' document.
But some of the entries there are outdated, some are questionable
and some more are simply not clear about the problem or the goal.
Some important entries are also missing.

So, remove what's needed to be removed, reformulate unclear
entries and add a bunch of new things that should be done.
Also, rename the file to 'TODO.md' as this express more clearly
the intent of the document.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>