add test case bug expand union
constant_symbol_value() needs several conditions before returning
a valid value.

However not all conditions are always met. One such case occurs when
an union is involved and the value we look after is of a different
type than the initializer.

Add two test cases showing (some of) the problems.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>