opcode: add OPF_TARGET
Quite a bit of code needs to know, in one way or another, if a
given instruction produces a result (insn->target) or not.

However, this information is not explicit and when needed it
must be retrieved with correct switch/case statements. These
statements need to be updated when a new instruction is added.

Now that there is a table containing some of the instructions
properties, add to this table a flag telling if the instruction
produces a result or not.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>