graph: do not use insn->symbol for memops
In struct instruction, OP_LOAD & OP_STORE are defined like unops
and so the pseudo holding the address should be 'insn->src'.
OTOH, 'insn->symbol' is supposed to be used only for OP_SYMADDR.

However, in graph.c, the field 'symbol' is used for OP_LOAD/STORE's
address. Now, this doesn't matter much as the field 'src' & 'symbol'
are aliased/unioned (together with 'src1', 'cond', ...) but it
can be confusing.

Change this by using the field 'src' for memops here too.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>