comparison: introduce impossibly_high_comparison() [fix build error]
This fixes a build error.  My local tests are supposed to prevent build
errors, but they were buggy in this case so they ended up testing the wrong
code.  I've updated them.

The story here is that when we check for index overflows, one thing we do
is we say: "Let's assume we overflowed, does that mean we have to assume
other impossible conditions are met?"  This uses comparison data to see if
the comparisons are still possible.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>