mem_tracker: add a --mem option to debug memory use problems
I found that Smatch was taking too long to test the kernel, and it was
because there were some places where we'd use all the memory and DoS the
build box.  Adding a --mem option gives us a way to find how much memory
we're using and which kernel .c files are the most tricky to parse.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>