ssa: phase 2: rename load & stores
This implement the second phase of the classical SSA conversion:
the renaming of variables but only for load & stores.

The implementation doesn't use the stack based algorithm (which
is quite hairy and can use a lot of memory) but a simpler solution
where each BB keep track of the symbol-to-pseudo correspondance.

This mapping is also done for phi-nodes (which are semantically
close to loads) but the phi-nodes themselves are not renamed.
This is done so that we can keep track of the unused phi-nodes
and so have semi-pruned SSA.

One thing worth of noticing is that in order to keep the mapping
for phi-nodes it was needed to add a field to OP_PHI instructions
to hold the original symbol pseudo (much like OP_LOAD).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>