ssa: phase 3: rename phi-nodes
This implement the last phase of the SSA conversion:
the renaming of the variables defined by a phi-nodes.

In this implementation, this need to be done last because
* we need the know the pseudo of each operands
* we place the phi-source lazily.

The implementation itself is very straight-forward and scan
after all OP_PHIs, create the phi-sources and simply use the
mapping of the preceding phase to lookup their value.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>