check_or_vs_and: ignore the kernel's min/max macros
I recently added support to parse __builtin_choose_expr() and that makes
most of these warnings go away but there are still about five files which
generate warnings.  Let's just silence these manually.

I thought about ignoring all macros and maybe that's the way to go in the
end, but for now this works in the kernel.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>