expressions: just allocate zero_expr() fresh every time
The problem here is that I was thinking about how zero->parent works.  We
presumably reset it often so the ->parent pointer can point to nonsense.
It's probably not a big deal, because of how things are normally parsed but
just from a correctness point of view, it seems like the right thing.

The other thing is that I changed it to &int_ctype.  It shouldn't make any
difference but it feels more correct.  Probably the most correct thing
would be to force people to specify the type...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>