math: fix a get_value() bug
The problem is that get_value() is supposed to get literals and constants.
But say we have a custom_handler set, like get_absolute_value() or
get_user_rl() then it does that instead.  And then to make things worse,
the answer gets cached so now everyone gets the wrong value when they call
get_value().

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>