default to LP64 for all and only for 64 bit ABIs
Currently sparse assumes LP64 only for x86-64.

This requires to always use the flag '-m64' for the other
64 bit archs like arm64 or ppc64.
There is also another problem on machines using the x32 ABI on
x86_64, which is, of course, not LP64.

So, change this by using LP64 by default if __LP64__ is defined
on the build machine. This will solve the current problem with
x32 but will also give the correct default on all 64 bit archs.

Reported-by: Uwe Kleine-König <uwe@kleine-koenig.org>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>