add a flag -mx32 ILP32 env on 64 bit archs
Currently, we have the -m32 & -m64 flags to force sparse to
behave as a 32 or 64 bit arch, independently of the arch used
to build sparse. Good.

However, on x86-64, '-m32' is there to force the 32 bit instructions
& ABI but there is another flag '-mx32' for the x32 ABI which use the
x86-64 instructions / 64 bit registers but where the pointers and
long ints are 32 bit.

Add the same flag, '-mx32', to support this ABI with the proper size
alignment for pointers and longs and to predefine __ILP32__ as it is
done by 'GCC -mx32'.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>
Acked-by: Ramsay Jones <ramsay@ramsayjones.plus.com>