testsuite: add check-assert
Ideally, the testcases should be universal but it happen than
some of them need to test some specificities or are meaningless
or plainly wrong in some situations. In such cases, these tests
must but ignored.

Currently, the only the only mechanisms a test are:
1) ignoring the tests depending on a tool which cannot be compiled
   (like, for example, those using sparse-llvm when LLVM is not
   installed.
2) some rather corse criteria using the name of the arch used
   to run the tests.

Allow more flexibility by allowing to exclude some tests based on
the success or failure of an arbitrary condition via _Static_assert().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>