enum: fix UB when rshifting by full width
Shifting by an amount greater or equal than the width
of the type is Undefined Behaviour. In the present case,
when type_is_ok() is called with a type as wide as an
ullong (64 bits here), the bounds are shifted by 64
which is UB and at execution (on x86) the value is simply
unchanged (since the shift is done with the amount modulo 63).

This, of course, doesn't give the expected result and
as consequence valid enums can have an invalid base type
(bad_ctype).

Fix this by doing the shift with a small helper which
return 0 if the amount is equal to the maximum width.

NB. Doing the shift in two steps could also be a solution,
  as maybe some clever trick, but since this code is in
  no way critical performance-wise, the solution here
  has the merit to be very explicit.

Fixes: b598c1d75a9c455c85a894172329941300fcfb9f
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>