enum: warn on bad enums
During the parsing of enum definitions, if some invalid
type combination is reached, the base type is forced to
'bad_ctype'. Good.

However, this is done without a warning and it's only when
the enum is used that some sign of a problem may appear,
with no hint toward the true cause.

Fix this by issuing a warning when the base type becomes invalid
but only if the type of the enumerator is itself not already
set to 'bad_ctype' (since it this case a more specific warning
has already been issued).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>