enum: only warn (once) when mixing bitwiseness
As an extension to the standard C types, parse supports bitwise
types (also called 'restricted') which should in no circonstances
mix with other types.

In the kernel, some enums are defined with such bitwise types
as initializers; the goal being to have slightly more strict enums.
While the semantic of such enums is not very clear, using a mix
of bitwise and not-bitwise initializers completely defeats the
desired stricter typing.

Attract some attention to such mixed initialization by issuing
a single warning for each such declarations.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>