enum: default to unsigned
GCC uses an unsigned type for enum's basetype unless one of
the enumerators is negative.

Using 'int' for plain simple enumerators and then using the same
rule as for integer constants (int -> unsigned int -> long -> ...)
should be more natural but doing so creates useless warnings when
using sparse on the kernel code.

So, do the same as GCC:
* uses the smaller type that fits all enumerators,
* uses at least int or unsigned int,
* uses an signed type only if one of the enumerators is negative.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>