enum: use the smallest type that fit
The C standard requires that the type of enum constants is 'int'
and let the enum base/compatible type be implementation defined.
For this base type, instead of 'int', GCC uses the smallest type
that can represent all the values of the enum (int, unsigned int,
long, ...)

Sparse has the same logic as GCC but if all the initializers
have the same type, this type is used instead.
This is a sensible choice but often gives differents
result than GCC.

To stay more compatible with GCC, always use the same logic
and thus only keep the common type as base type for restricted
types.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>