flow: don't parse inline functions which aren't interesting
One thing which is a problem is that inline functions generate tons and
tons of duplicate warnings, but the main thing here that I was worried
about was that we were filling the DB with nonsense.

In the original code, we would parse inline and because they are static
and there are no callers in this file then we'd record that "This function
is never called".  So then the inline calls other functions with
"everything is unknown".  Then the other callers would sometimes be like,
"Ugh...  Too much information from too many callers.  Let's ignore it all".

So now if we have an uncalled inline in a .c file then we say "Ok.  This
is legit an uncalled inline and everything is unknown".  But otherwise we
don't even parse it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>