db/untracked_param: mark parameters untracked if we have too much info
There are some functions that when we parse them there is just too much
information so we throw our hands up and mark that "We return here" but we
don't record anything else.  It sometimes causes "variable uninitialized
bugs because there isn't a PARAM_SET in the return data".  In that case,
what we're supposed to do is mark the parameters as untracked.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>