function_ptrs: record local pointer scope correctly
Say you have a function pointer that's local to a specific function, then
that's supposed to be recorded as "function_name pointer_name".  But the
is_local_symbol() function always returned false so everything was treated
as global.

And then because there are a bunch of function pointer called "func", it
meant they were all mixed together.  Which was a mess because it meant
that about a 100 functions were called with too much data and we just
threw it all away.  And probably there were some function which were called
with not such common pointer names and so it was mixed with invalid data.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>