extra: handle comparisons to expression statements
I've been running this code for a while and I forget some of the details.
It probably has to do with READ_ONCE().

If (foo > READ_ONCE(bar)) {, then the READ_ONCE() macro expands to
something like:

({
	...
	...
	__u.value = bar;  <-- faked by smatch
	__u.value;
})

So we want to parse that all and fake a comparison without the intermediary
steps so it looks like "if (foo > bar) {".

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>