db: make call_implies rows unique
There was an issue with recursion a while back so call_implies kept
growing exponentially in commit f2fc154233fb ("about_fn_ptr_arg: don't get
into a recursive loop").

That made me think that probably call implies lines should be unique
anyway.

I wish I had tested this to make sure it doesn't cause a slow down when you
build the DB...  I've been running with this change for a while and building
my DB is slow but my code has a bunch of other changes as well so it's
probably something else causing the slow down.  I'm not comfortable just
carrying it in my tree in case I miss a regression in the published code.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>