doc: automatically get the version
Documentation generated by sphinx reference the version of the
project being documented. This version need to be specified via
sphinx's config file.

However, if this version number is simply added there, it's
guaranteed that soon or later it will be out-of-synch with git's
version tag (or the version number present in the Makefile).

Fix this by trying to read the version from git or, if this fails,
from the Makefile.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>