doc: add sphinx domain for IR instruction indexation
The doc for IR instructions used definition lists to structure
the instructions names and their description. This is simple
to write and present well but it doesn't put the instructions'
name in the index.

Fix this by adding a new domain, with a single directive 'op'
which takes the instruction name, inserts an index entry and uses
it for the definition.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>